In the Admin Panel, unlink the last item in the breadcrumbs trail if it would link to the current page. This should fix bug [#810] Breadcrumbs don't work for the help pages. Links back to the same page.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@3117 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/themes/default/defaultTheme.php
@@ -102,9 +102,16 @@ class defaultTheme extends AdminTheme
 			if ($counter > 0) {
 				echo " &#187; ";
 			}
-			echo '<a class="breadcrumbs" href="'.$crumb['url'];
-            echo '">'.$crumb['title'];
-            echo '</a>';
+			if (str_replace('&amp;', '&', $crumb['url']) != basename($_SERVER['REQUEST_URI']))
+			  {
+			    echo '<a class="breadcrumbs" href="'.$crumb['url'];
+			    echo '">'.$crumb['title'];
+			    echo '</a>';
+			  }
+			else
+			  {
+			    echo $crumb['title'];
+			  }
 			$counter++;
 		}
